v0.7[deadline]

+++ Assets/Scripts/System/LevelControl/BuilderDirector.cs
@@ -14,53 +14,71 @@ public class BuilderDirector : MonoBehaviour
     public Image background;
     public int time = 0;
     public int score = 0;
+    public int levelIndex;
     public List<GameObject> enemies = new List<GameObject>();
     public GameObject player;
+    public bool TestingState;
 
     private void Start()
     {
         //builder.level = builder.BuildLevel(GlobalData.levelName);
-        //builder.BuildPlayer(new PlayerModel(new Vector3(0, 0, 0), new Quaternion(0, 0, 0, 0), new Vector3(10, 10, 0), false, false, true, 2000, 2000, new List<Vector2>() { new Vector2(0.3f, 0), new Vector2(-0.3f, 0) }, new List<float>() {0,0 }, 2, 0.1f), ref cam);
+        //builder.BuildPlayer(new PlayerModel(new Vector3(0, 0, 0), new Quaternion(0, 0, 0, 0), new Vector3(10, 10, 0), false, false, true, 2000, 2000, new List<Vector2>() { new Vector2(0.3f, 0), new Vector2(-0.3f, 0) }, new List<float>() { 0, 0 }, new List<float>() {0.5f, 0.5f }, 2, 0.1f), ref cam);
         //builder.BuildMeleeEnemy(new MeleeModel(new Vector3(80, 80, 0), new Quaternion(0, 0, 0, 0), new Vector3(10, 10, 0), false, false, false, 1000, 1500, 1, 30));
-        //builder.BuildRangeEnemy(new RangeModel(new Vector3(70, 70, 0), new Quaternion(0, 0, 0, 0), new Vector3(10, 10, 0), false, false, false, 1500, 1500, new List<Vector2>() { new Vector2(0.3f, 0f), new Vector2(-0.3f, 0) }, new List<float>() { 0, 0 }, 2, 0.5f, 40, 50));
+        //builder.BuildRangeEnemy(new RangeModel(new Vector3(70, 70, 0), new Quaternion(0, 0, 0, 0), new Vector3(10, 10, 0), false, false, false, 1500, 1500, new List<Vector2>() { new Vector2(0.3f, 0f), new Vector2(-0.3f, 0) }, new List<float>() { 0, 0 }, new List<float>() { 0.5f, 0.5f }, 2, 0.5f, 40, 50));
         //builder.BuildWall(new WallModel(new Vector3(25, 25, 0), new Quaternion(0, 0, 0, 0), new Vector3(1, 15, 0)));
         //decomposer.levelModel.backgroundSpritePath = "Textures\\Background";
         //background.sprite = Resources.Load<Sprite>(decomposer.levelModel.backgroundSpritePath);
         //builder.BuildPrefab("Prefabs\\Units\\Boss");
-      
+
         builder.level = builder.BuildLevel(GlobalData.levelName);
-        if (File.Exists(builder.level.LevelName))
+        if(!TestingState)
         {
-            Serialiazer.DeserialiazitionFromXml(ref decomposer.levelModel, builder.level.LevelName);
-            foreach (WallModel item in decomposer.levelModel.wallList)
+            if (File.Exists(builder.level.LevelName))
             {
-                builder.BuildWall(item);
-            }
-            foreach (MeleeModel item in decomposer.levelModel.meleeList)
-            {
-                builder.BuildMeleeEnemy(item);
-            }
-            foreach (RangeModel item in decomposer.levelModel.rangeList)
-            {
-                builder.BuildRangeEnemy(item);
-            }
-            builder.BuildPlayer(decomposer.levelModel.playerObj, ref cam);
-            background.sprite = Resources.Load<Sprite>(decomposer.levelModel.backgroundSpritePath);
-            time = decomposer.levelModel.levelTime;
-            score = decomposer.levelModel.levelScore;
+                Serialiazer.DeserialiazitionFromXml(ref decomposer.levelModel, builder.level.LevelName);
+                if (GlobalData.levelIndex == -1)
+                {
+                    GlobalData.levelIndex = decomposer.levelModel.levelIndex;
+                }
+                foreach (WallModel item in decomposer.levelModel.wallList)
+                {
+                    builder.BuildWall(item);
+                }
+                foreach (MeleeModel item in decomposer.levelModel.meleeList)
+                {
+                    builder.BuildMeleeEnemy(item);
+                }
+                foreach (RangeModel item in decomposer.levelModel.rangeList)
+                {
+                    builder.BuildRangeEnemy(item);
+                }
+                builder.BuildPlayer(decomposer.levelModel.playerObj, ref cam);
+                background.sprite = Resources.Load<Sprite>(decomposer.levelModel.backgroundSpritePath);
+                time = decomposer.levelModel.levelTime;
+                score = decomposer.levelModel.levelScore;
 
-            player = builder.level.LevelObjects.Find(o => o.tag == "Player");
-            enemies.AddRange(builder.level.LevelObjects.Where(o => o.tag.Contains("Enemy")));
+                player = builder.level.LevelObjects.Find(o => o.tag == "Player");
+                enemies.AddRange(builder.level.LevelObjects.Where(o => o.tag.Contains("Enemy")));
 
-            decomposer.levelModel.meleeList.Clear();
-            decomposer.levelModel.rangeList.Clear();
-            decomposer.levelModel.wallList.Clear();
-            decomposer.levelModel.playerObj = null;
+                decomposer.levelModel.meleeList.Clear();
+                decomposer.levelModel.rangeList.Clear();
+                decomposer.levelModel.wallList.Clear();
+                decomposer.levelModel.playerObj = null;
+            }
+            else
+            {
+                SceneManager.LoadSceneAsync("MainMenu");
+            }
         }
         else
         {
-            SceneManager.LoadSceneAsync("MainMenu");
+            builder.BuildPlayer(new RangeModel(new Vector3(0, 0, 0), new Quaternion(0, 0, 0, 0), new Vector3(10, 10, 0), false, false, true, 1000, 1000, new List<Vector2>(), new List<float>(), new List<float>(), 0, 0.5f, 0, 0, 0.1f, 100, GlobalData.prefabBulletLazer),ref cam);
         }
+    }
+
+    private void Update()
+    {
+        builder.level.LevelObjects.ForEach(o => { if (o == null) { builder.level.LevelObjects.Remove(o); enemies.Remove(o); } });
     }
 
     //public void clll()
@@ -71,14 +89,13 @@ public class BuilderDirector : MonoBehaviour
 
     public void SaveGame(string path)
     {
-        //Object[] q;
-        //q = FindSceneObjectsOfType(typeof(GameObject));
-        //for(int i = 0;i< q.Length;i++)
-        //Debug.Log(q.GetValue(i));
-
-        decomposer.levelModel.levelScore = score;
-        decomposer.levelModel.levelTime = time;
-        File.Delete(path);
+        if(!TestingState)
+        {
+            decomposer.levelModel.levelScore = score;
+            decomposer.levelModel.levelTime = time;
+            decomposer.levelModel.levelIndex = GlobalData.levelIndex;
+            File.Delete(path);
+        }
         foreach (GameObject item in builder.level.LevelObjects)
         {
             if (item != null)
@@ -103,7 +120,40 @@ public class BuilderDirector : MonoBehaviour
         }
        
         Serialiazer.SerialiazeToXml(ref decomposer.levelModel, path);
+        decomposer.levelModel.meleeList.Clear();
+        decomposer.levelModel.rangeList.Clear();
+        decomposer.levelModel.wallList.Clear();
+        decomposer.levelModel.playerObj = null;
     }
+
+    #region Testing mode
+
+    public void AddWall()
+    {
+        builder.BuildWall(new WallModel(new Vector3(0, 0, 0), new Quaternion(0, 0, 0, 0), new Vector3(1, 1, 0)));
+    }
+
+    public void AddMelee()
+    {
+        builder.BuildMeleeEnemy(new MeleeModel(new Vector3(0, 0, 0), new Quaternion(0, 0, 0, 0), new Vector3(10, 10, 0), false, false, false, 1000, 1000, 0, 0, 0.1f, 100));
+    }
+
+    public void AddRange()
+    {
+        builder.BuildRangeEnemy(new RangeModel(new Vector3(0, 0, 0), new Quaternion(0, 0, 0, 0), new Vector3(10, 10, 0), false, false, false, 1000, 1000, new List<Vector2>(), new List<float>(), new List<float>(), 0, 0.5f, 0, 0, 0.1f, 100, GlobalData.prefabBulletLazer));
+    }
+
+    public void AddPrefab(string prefab)
+    {
+        builder.BuildPrefab(prefab);
+    }
+
+    public void BackToMainMenu()
+    {
+        SceneManager.LoadScene("MainMenu");
+    }
+
+    #endregion
 
     private void OnDestroy()
     {
